fix: M4b D5.1 — address adversarial review of #810: probe decoupled from task completion (own starting-surface project), final-build gating kills stale-DLL false catch, W5B-003 renamed off the invalid-run marker, smoke integrity enforced (hook both arms + tamper hash), C#-arm scratch tests unforeclosed, spec wording no longer orders preserving the defect, parity convention normalized
