+++ bench/phase0-agent-native/pairs/W5A-001-report-audit/defect.json
@@ -5,7 +5,7 @@
   "description": "FormatSummary is declared pure (calor: §E{}; csharp: doc comment 'Pure: no side effects') but appends to a relative-path audit log via a named static call (File.AppendAllText -> fs:w per the bcl-io manifest).",
   "coveredPath": "named static BCL call inside a first-order function body — EffectEnforcementPass manifest resolution; no delegates, no dispatch (the excluded holes are not involved)",
   "calorCatchChannel": "build-blocking Calor0410 with call chain (FormatSummary -> File.AppendAllText), fires on the agent's first build",
-  "probeTest": "ReportPair.HeldOut.ReportHeldOutTests.Probe_FormatSummary_PerformsNoFileWrite",
+  "probeTest": "ReportPair.Probe.ProbeTests.Probe_FormatSummary_PerformsNoFileWrite",
   "acceptance": {
     "calorFixtureBlocksAtBuild": true,
     "csharpFixtureBuildsGreen": true,
