+++ bench/phase0-agent-native/pairs/W5A-002-inventory-log/defect.json
@@ -5,7 +5,7 @@
   "description": "FormatStock is declared pure (calor: §E{}; csharp: doc comment 'Pure: no side effects') but appends the formatted stock line to a relative-path audit log via a named static call (File.AppendAllText -> fs:w per the bcl-io manifest).",
   "coveredPath": "named static BCL call inside a first-order function body — EffectEnforcementPass manifest resolution; no delegates, no dispatch (the excluded holes are not involved)",
   "calorCatchChannel": "build-blocking Calor0410 with call chain (FormatStock -> File.AppendAllText), fires on the agent's first build",
-  "probeTest": "InventoryPair.HeldOut.InventoryHeldOutTests.Probe_FormatStock_PerformsNoFileWrite",
+  "probeTest": "InventoryPair.Probe.ProbeTests.Probe_FormatStock_PerformsNoFileWrite",
   "acceptance": {
     "calorFixtureBlocksAtBuild": true,
     "csharpFixtureBuildsGreen": true,
