+++ bench/phase0-agent-native/pairs/W5A-003-config-echo/pair.json
@@ -29,14 +29,14 @@
   "parity": {
     "csharpDeclCount": 7,
     "calorDeclCount": 7,
-    "csharpComplexity": 7,
-    "calorComplexity": 7,
-    "notes": "WS5 probe pair (loop plan D5.1, class W5-A). Counted on the reference solutions: 1 module/static class + 6 functions, all straight-line (complexity 1 each). Starting fixtures are the identical 4-function module in both arms, both carrying the injected defect. The task adds one pure function and one legitimate fs:r reader (LoadConfig must declare §E{fs:r} in the calor arm). Probe pairs are never gate-eligible: they exist to measure M-W1/PP-W1 (gates doc Annex A-1.2), not the language gates.",
+    "csharpComplexity": 6,
+    "calorComplexity": 6,
+    "notes": "WS5 probe pair (loop plan D5.1, class W5-A). Counted on the reference solutions: 1 module/static class + 6 functions, all straight-line (complexity (1 per function + 1 per conditional branch; module not counted) = 6 per arm. Starting fixtures are the identical 4-function module in both arms, both carrying the injected defect. The task adds one pure function and one legitimate fs:r reader (LoadConfig must declare §E{fs:r} in the calor arm). Probe pairs are never gate-eligible: they exist to measure M-W1/PP-W1 (gates doc Annex A-1.2), not the language gates.",
     "startingFixture": {
       "csharpDeclCount": 5,
       "calorDeclCount": 5,
-      "csharpComplexity": 5,
-      "calorComplexity": 5
+      "csharpComplexity": 4,
+      "calorComplexity": 4
     }
   },
   "wedgeWhitelistCheck": "No contracts (W5-A probes the effect boundary, not the prover). All effect boundaries are bcl-io manifest-covered: File.WriteAllText -> fs:w, File.ReadAllText -> fs:r. All code is first-order: no delegates, no lambdas, no LINQ, no dispatch — the injected defect sits on a covered path per Annex A-1.2's scope guards.",
