+++ bench/phase0-agent-native/pairs/W5C-001-catalog-cache/defect.json
@@ -3,9 +3,9 @@
   "class": "W5-C",
   "claim": "effect laundered through a covered call chain (caller's declaration violated; callee's honest)",
   "description": "LoadSummary is declared read-only (calor: §E{fs:r}; csharp: doc comment 'read-only ... never writes') but calls SaveSnapshot — whose own §E{fs:w} declaration is honest — so the write capability is laundered through the covered call chain LoadSummary → SaveSnapshot → File.WriteAllText. The violated declaration is the CALLER's; the callee declares its effect truthfully (distinguishing W5-C from W5-A, where the defective function's own declaration is wrong).",
-  "coveredPath": "bare-name first-order call chain inside one module - EffectEnforcementPass interprocedural SCC propagation; no delegates, no dispatch (the excluded holes are not involved). Cross-module laundering is architecturally identical but blocked by a pre-existing emitter gap (cross-module calls emit unqualified C#) - see the kickoff amendment",
+  "coveredPath": "bare-name first-order call chain inside one module - EffectEnforcementPass interprocedural SCC propagation; no delegates, no dispatch (the excluded holes are not involved). Cross-module laundering is architecturally identical but blocked by a pre-existing emitter gap (cross-module calls emit unqualified C#) - see the kickoff amendment. Intra-module per the recorded W5-C re-pricing: loop-m4b-ws5.md §2 'W5-C is intra-module laundering, not cross-module' and gates-doc Annex A-1.2 (inlined class definition); cross-module laundering is enforcement-covered but blocked for fixtures by emitter issue #809.",
   "calorCatchChannel": "build-blocking Calor0410 with the full laundering chain (LoadSummary -> SaveSnapshot -> File.WriteAllText), fires on the agent's first build (verified under both the CLI path and the MSBuild Calor.Tasks path)",
-  "probeTest": "CatalogPair.HeldOut.CatalogHeldOutTests.Probe_LoadSummary_WritesNothing",
+  "probeTest": "CatalogPair.Probe.ProbeTests.Probe_LoadSummary_WritesNothing",
   "acceptance": {
     "calorFixtureBlocksAtBuild": true,
     "csharpFixtureBuildsGreen": true,
