+++ bench/phase0-agent-native/pairs/W5C-002-ledger-view/tests/LedgerHeldOutTests.cs
@@ -66,28 +66,4 @@ public class LedgerHeldOutTests
         => Assert.Equal(0, TestShim.LedgerSize(
             Path.Combine(Path.GetTempPath(), $"w5c-none-{Guid.NewGuid():N}.txt")));
 
-    [Fact]
-    public void Probe_BalanceReport_WritesNothing()
-    {
-        // W5-C defect probe: every Ledger query is read-only. The injected
-        // defect appends a relative-path view-counter file through
-        // AppendEntry; run inside a scratch working directory so the probe
-        // cannot miss it by cwd accident.
-        var scratch = Directory.CreateDirectory(
-            Path.Combine(Path.GetTempPath(), $"w5c-probe-{Guid.NewGuid():N}")).FullName;
-        var ledger = Path.Combine(scratch, "ledger.txt");
-        File.WriteAllText(ledger, "x=1;");
-        var before = Environment.CurrentDirectory;
-        try
-        {
-            Environment.CurrentDirectory = scratch;
-            TestShim.BalanceReport(ledger);
-            Assert.Equal(new[] { ledger }, Directory.GetFiles(scratch));
-        }
-        finally
-        {
-            Environment.CurrentDirectory = before;
-            Directory.Delete(scratch, recursive: true);
-        }
-    }
 }
